[#116538279] db migrate on test got into a state where the rake file was registered as run but the column had been removed. Added a check to the migration scripts to prevent that from breaking.

+++ db/migrate/20160429133315_remove_unused_columns_from_disposition_update_events.rb
@@ -1,5 +1,7 @@
 class RemoveUnusedColumnsFromDispositionUpdateEvents < ActiveRecord::Migration
   def change
-    remove_column :asset_events, :new_owner_name, :string, :limit => 64
+    if column_exists? :asset_events, :new_owner_name
+      remove_column :asset_events, :new_owner_name, :string, :limit => 64
+    end
   end
 end
